Move mocks to their respective packages to avoid circular dependencies, add Referee
