+++ game/game_status.go
@@ -1,0 +1,9 @@
+package game
+
+type GameStatus string
+
+const (
+	Win      GameStatus = "win"
+	Tie      GameStatus = "tie"
+	Continue GameStatus = "continue"
+)
