+++ game/game_test.go
@@ -3,7 +3,7 @@ package game
 import "github.com/stretchr/testify/assert"
 import . "github.com/go-number-1-fan/tic-tac-toe/board"
 import . "github.com/go-number-1-fan/tic-tac-toe/player"
-import . "github.com/go-number-1-fan/tic-tac-toe/mocks"
+import . "github.com/go-number-1-fan/tic-tac-toe/ui"
 import "testing"
 
 func TestGame_HasABoard(t *testing.T) {
