+++ game/referee.go
@@ -1,0 +1,7 @@
+package game
+
+import . "github.com/go-number-1-fan/tic-tac-toe/board"
+
+type Referee interface {
+	GetGameStatus(board Board) GameStatus
+}
