+++ player/mock_player.go
@@ -1,4 +1,4 @@
-package mocks
+package player
 
 import . "github.com/go-number-1-fan/tic-tac-toe/board"
 import . "github.com/go-number-1-fan/tic-tac-toe/ui"
