+++ ui/mock_ui.go
@@ -1,4 +1,4 @@
-package mocks
+package ui
 
 import . "github.com/go-number-1-fan/tic-tac-toe/board"
